Add tests for view query
