+++ test/test_helper.rb
@@ -13,6 +13,7 @@ Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
 class Test::Unit::TestCase
   DEFAULT_DB = 'test'
+  DB_BASE_PATH = "http://localhost:5984/test".freeze
 
   setup :global_setup
 
@@ -27,6 +28,51 @@ class Test::Unit::TestCase
 
   def base_path
     %r{http://.*:5984/#{DEFAULT_DB}}
+  end
+
+  def generic_response rows, count = 1
+    {total_rows: count, offset:0, rows: rows}
+  end
+
+  def build_view_response properties
+    rows = properties.map.with_index do |v, i|
+      {
+        id: "foo_bar/#{i}",
+        key: "foo_bar",
+        value: 1,
+        doc: {_id: "foo_bar/#{i}", _rev: SecureRandom.hex}.merge!(v)
+      }
+    end
+    generic_response rows, properties.count
+  end
+
+  def build_view_collation_response properties
+    rows = properties.map.with_index do |v, i|
+      id = i.zero? ? "foo_bar/#{i}" : "baz/#{i}"
+      {
+        id: id,
+        key: "foo_bar",
+        value: 1,
+        doc: {_id: id, _rev: SecureRandom.hex}.merge!(v)
+      }
+    end
+    generic_response rows, properties.count
+  end
+
+
+  def build_request keys, body, view_name = 'foo_bar'
+    query = "keys=#{CGI::escape keys.to_s.gsub(' ','')}&" unless keys&.empty?
+    stub_request(:get, "#{query_base_path}?#{query.to_s}include_docs=true").
+      to_return(body: body.to_json)
+  end
+
+  def query_base_path
+    "#{DB_BASE_PATH}/_all_docs"
+  end
+
+  def build_save_request(obj)
+    stub_request(:put, "#{DB_BASE_PATH}/#{CGI.escape(obj.id)}").
+      to_return(body: {ok: true, id: obj.id, rev: "FF0000" }.to_json)
   end
 end
